+++ mokakbob-api/src/main/java/com/mokakbob/matching/service/MatchingStartService.java
@@ -2,36 +2,23 @@ package com.mokakbob.matching.service;
 
 import com.mokakbob.common.exception.exceptions.ApiException;
 import com.mokakbob.domain.matching.domain.vo.MatchingCategory;
-import com.mokakbob.domain.matching.service.MatchingService;
-import com.mokakbob.domain.member.domain.Member;
-import com.mokakbob.domain.member.service.MemberService;
 import com.mokakbob.matching.exception.MatchingErrorCode;
-import com.mokakbob.matching.ParticipantRedisStore;
-import com.mokakbob.matching.service.event.MatchingParticipateEvent;
-import java.math.BigDecimal;
 import java.util.concurrent.TimeUnit;
 import lombok.RequiredArgsConstructor;
 import org.redisson.api.RLock;
 import org.redisson.api.RedissonClient;
-import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
 public class MatchingStartService {
 
-    private static final int DEFAULT_DEDUCE_POINT = 2000;
     private static final int LIMIT_LOCK_CATCH_TIME = 3;
     private static final int LOCK_DURATION_TIME = 10;
 
-    private final ParticipantRedisStore participantStore;
-    private final MemberService memberService;
-    private final MatchingService matchingService;
-    private final ApplicationEventPublisher eventPublisher;
+    private final MatchingTransactionService matchingTransactionService;
     private final RedissonClient redissonClient;
 
-    @Transactional
     public void participateMatchingWithLock(double lat, double lng, MatchingCategory category, int participantCount,
                                             Long memberId) {
         String lockKey = "lock:member:" + memberId + ":participation";
@@ -39,7 +26,7 @@ public class MatchingStartService {
 
         try {
             if (lock.tryLock(LIMIT_LOCK_CATCH_TIME, LOCK_DURATION_TIME, TimeUnit.SECONDS)) {
-                participateMatching(lat, lng, category, participantCount, memberId);
+                matchingTransactionService.participateMatching(lat, lng, category, participantCount, memberId);
             } else {
                 throw new ApiException(MatchingErrorCode.EXIST_MATCHING);
             }
@@ -50,36 +37,6 @@ public class MatchingStartService {
             if (lock.isHeldByCurrentThread()) {
                 lock.unlock();
             }
-        }
-    }
-
-    private void participateMatching(double lat, double lng, MatchingCategory category, int participantCount,
-                                     Long memberId) {
-        validateExistParticipating(memberId);
-        deducePoint(memberId);
-
-        matchingService.saveMatchingRequest(memberId, category, participantCount, new BigDecimal(lat),
-                new BigDecimal(lng));
-
-        MatchingParticipateEvent event = new MatchingParticipateEvent(memberId, lat, lng, category, participantCount);
-
-        eventPublisher.publishEvent(event);
-    }
-
-
-    private void deducePoint(Long memberId) {
-        Member member = memberService.findMember(memberId);
-
-        if(member.getDepositPoint() < DEFAULT_DEDUCE_POINT) {
-            throw new ApiException(MatchingErrorCode.NOT_ENOUGH_MATCHING_POINT);
-        }
-
-        member.deductPoint(DEFAULT_DEDUCE_POINT);
-    }
-
-    private void validateExistParticipating(Long memberId) {
-        if (participantStore.isAlreadyParticipating(memberId)) {
-            throw new ApiException(MatchingErrorCode.EXIST_MATCHING);
         }
     }
 }
